SIP Qualify Support
Completed the work that had been started for SIP qualify support in the new
channel driver.  Qualify requests are now sent for every contact associated
with an endpoint.  When a contact is registered a check is done to make sure
it is available and the round trip time is stored for reference.  The contact
continues to be checked at a specified interval (qualify_frequency - now set
on the AOR) until it is removed.

A CLI command ("sip qualify <endpoint>") is also available that when issued
will send qualifies for all contacts on the given endpoint.  Also added CLI
command "sip show endpoint <endpoint>" that will show all the contacts and
their status for the given endpoint.

(closes issue ASTERISK-21499)
Reported by: Matt Jordan
Review: https://reviewboard.asterisk.org/r/2584/