Fix an assert when a partially handled SUBSCRIBE was rejected.
What was happening is the SUBSCRIBE was being handled and a dialog (and transaction)
created. When trying to send a 500 back statelessly the code asserted as it expects
no transaction to exist.

This change terminates the subscription on failure cases and also sends the 500 response
statefully when needed. Once the response is sent and the timer ends everything is
properly destroyed.