Get rid of a XXX comment in the COMPLETE state.
There was a concern about sending an UNHOLD to a bridge
after a merge since either of the individual bridges may have
been simple bridges, but the merge may have resulted in the
bridge being a multimix bridge. If an UNHOLD is sent to a multimix
bridge, then the frame will not be passed to the channels.

This is corrected by performing the unhold before entering the
COMPLETE state. Since there are separate states depending on
what bridge the transferer is in prior to the COMPLETE state, this
makes things very simple.