Get blond transfers working properly.
This change involves changing the BLOND state to a terminal
one. The idea originally was for this state to not be terminal
so that it could transition in case we need to re-call any of
the parties involved. However, I am instead going to create
a different state to be used for that scenario.