Some progress towards getting recalls more stable.
This divides the state machine into two superstates,
transferring, and recalling. The transferring state is
already stable and involves the typical transfer mechanics.
The recalling state is the one that is unstable at the moment,
and involves the logic for re-calling the transferer and
transfer target when atxferdropcall=no.

As part of this conversion, the pull method for atxfer bridges
has been divided in two. The logic is different depending on
which superstate the transfer is currently in.

This commit also contains a fix for a problem I would see sometimes
where a hangup hook on a transferer would trigger twice. The reason
was because when I detected a hangup on the transferer channel, I
would not include his bridge channel in the list of kicked channels
when merging bridges. By doing this, it could cause the hangup hook
to be triggered initially when the transferer hung up and again after
the transferer was pushed into the merge target. I now always include
the transferer's channel as a channel to kick when merging bridges because
of a completed or blond transfer. However, I have to check that the
transferer channel's bridge channel is not NULL or else I may trigger
a crash when attempting to merge.