Change certain transitions between CONSULTING and DOUBLECHECKING.
Rather than putting conditional logic in the enter callbacks for these
states, I instead placed the logic when transitioning to these states.
It's easier to make sure that preconditions are satisfied from the previous
state than to not have preconditions and try to guess the current conditions
in the new state.