Add timert1 and timerb options.
These options are a bit different from others so far in that they affect how
PJSIP operates rather than how Asterisk operates. These, plus other PJSIP options
need to be set early and then cannot be changed due to a reload. The only way these
could be changed is if res_sip.so is unloaded and the loaded again.

Since these options cannot be reloaded, they need to be handled by a separate sorcery
than the rest of the SIP configuration. I've started by calling the sorcery as well
as the type for this configuration "system". I could have gone with "pjsip" or something
similar, but I think "system" gives us a place to put other non-pjsip related options that
we may introduce that also cannot be affected by reloads.