Add "useragent" and "maxforwards" options.
For these options, a new "global" section was created. This is similar
to the "system" section except that reloads will actually pick up new
values from the "global" section.

A new PJSIP module was added in order to facilitate these options. The
module registers handlers that are called on outgoing requests and responses
and adds the configured headers.