Commit some progress towards transfer detection.
With this change, I subscribe to stasis messages for bridges
when a caller enters a queue. I detect when the caller enters
the bridge and then hold onto that bridge for future reference
in case it is involved in a blind or attended transfer.

What I don't have in place is the appropriate queue log or stasis
publications when this occurs. That will be coming next. I've started
to modify the send_agent_complete() function but it isn't actually
used anywhere yet.