Resolve the final BUGBUG.
With this out of the way, there are no more BUGBUGs in app_queue.
However, there is still work to do to ensure we report events accurately
once a bridge is established.

Local channel optimization is the biggest hurdle. Either of the caller or
agent channels may be optimized and they may be optimized multiple times.
The way that local channel optimization works now also means that the bridge
we create in app_queue may not be the bridge that the caller and agent end
up actually being in.

The other hurdle, much smaller, is that currently internal attended transfers
that are completed with a hangup are being logged as hangups rather than as
transfers. This will be easy to resolve, but my plan for how to resolve it
unfortunately requires the atxfer_features branch to be merged.