Re-rework local channel optimization handling.
I realized that because of the fact that multiple local
channel optimizations can be reported to me at once, and the
events for local channel optimizations occur after they have
already happened, it simply is not possible for me to get
an accurate picture of all channels involved.

The remedy for this problem is to restrict what types of local
channel optimization are permitted when using queues. The only
type of local channel optimization that is allowed is move-swap
optimization and it's only allowed to happen towards the queue
bridge. This means that the following types of local channels
will not be able to optimize out:

* A local channel between a queue and a multi-party bridge
* A local channel between two queues.

Luckily, this still means that an estimated 99.999% of queue
usage will still have local channel optimizations work properly.
In the cases where the local channels do not optimize away, the
calls will still work, but there may be more channels in use
than expected. Tough luck.