Fix crash due to trying to send a re-invite while in the incorrect state.
This crash would occur if a re-invite was queued while the initial INVITE
transaction was still occurring and the response to the INVITE was not ACKed.
This lack of ACK would cause the INVITE session state to never reach confirmed.
Once the transaction terminated, however, the queued re-invite would occur and
cause a crash due to this lack of state change.

This fix checks the INVITE session state before performing the re-invite to
ensure it is in the required confirmed state.
........

Merged revisions 395455 from file:///srv/subversion/repos/asterisk/trunk
........

Merged revisions 395463 from file:///srv/subversion/repos/asterisk/team/rmudgett/bridge_phase