Un-anonymize the structs within an ast_sip_endpoint (except media).
This allows to easily create initialization/destruction functions for the various types,
should they be required. This in turn allows for some string parameters to be moved from
the top-level endpoint structure to places where they belong.

Media will be handled in its own commit since it is a pretty large changeset.