Create a branch in order to create an AMI message indicating a parked channel has been swapped.
When performing an attended transfer to a parking lot, the ParkedCall AMI event will show that
the transferer has been parked. This is because the transferer will be the channel that actually
runs the Park() application. Once the transfer completes, the transferee will swap into the
transferer's spot in the parking lot.

AMI messages, as of now, can actually be used to piece together that this has occurred, but it
requires more state to be kept by the consumer of the AMI messages than is desired. This branch
seeks to do two things to help out:

1) BridgeEnter events will have a second channel in them if the channel that is entering the
bridge is designated to swap a specific channel out of the bridge.

2) Parking will emit an event stating that a channel that is parked in a specific lot and
a specific space has been replaced with a different channel.

(issue ASTERISK-22193)