Fix how contact is generated for an AOR.
The host value was being copied verbatim. However, a full
SIP URI is expected for the contact. This means having to
construct a URI based on the username, host, and potentially
the port.

This also corrects an error where I had the parameters in the
wrong order for subscribemwi lookup.