Cache the string representation of channel nativeformats when nativeformats are set.
Channel snapshot creation was taking quite a bit of time iterating through formats in
order to get the string representation of the formats. Switching to copying formats had
no noticeable positive effect on performance. However, nativeformats are not set on a
channel very often. So if we cache the string representation at that point, then it becomes
trivial to copy this value onto the snapshot when the snapshot is created.