Fix IAX2 incoming call address lookups
This fixes address lookup for incoming calls without a peer definition.
The address family was unset instead of being set to AST_AF_UNSPEC
which was causing lookup failures on "127.0.0.1". This is one of the
causes of the current failure of the app_page integration test.

Review: https://reviewboard.asterisk.org/r/2933/